requirements.txt
flask
ollama
streamlit
langchain
google-generativeai
langchain-google-genai
